new approach: aligning images to coordinate grid
instead of the old approach, where template matching was used to match every single image, now, you can find where zero-zero is on each map automatically and beforehand. this method is a lot faster since it doesn't have to do template matching when cropping -- just looking up the offsets from a json file and using that to find a location on the map.